Makefile.am: Remove libopkg/config.h from distribution archive
A new target is added, opkg-dist, which creates a .tar.gz distribution archive
without the file libopkg/config.h.

This is a bit of a hack and is a temporary solution. The real solution will be
to ensure config.h is not installed and not needed by public API headers.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>

git-svn-id: http://opkg.googlecode.com/svn/branches/opkg-0.2.x@663 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358