Tidy up const usage around str_to_constraint function
The prototype of str_to_constraint is changed to reflect the fact that it
doesn't modify the strings it's given.

The prototype of parseDepends is changed to reflect the fact that it doesn't
modify the string it's given. This is needed to match the new prototype of
str_to_constraint

Usage of str_to_constraint in pkg_version_satisfied no longer needs to lose the
const qualifier.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>

git-svn-id: http://opkg.googlecode.com/svn/trunk@682 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358