opkg_conf: Zero all integer and bool options on deinit
String options are all freed and set to NULL, we should also ensure that integer
and boolean options are set to zero so that we don't run into issues if libopkg
is reinitialised by a call to opkg_conf_init() after the call to
opkg_conf_deinit().

This fixes a warning message issue when reusing the library context which
happens for non-string conf-file options (ie boolean and integer).

Signed-off-by: Gaël PORTAY <g.portay@overkiz.com>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>