pkg_hash.c: Handle hold and prefer flags separately
The previous fix to pkg_hash_fetch_best_installation_candidate improved things
but didn't fix everything. If a package filename is specified on the command
line it gets the SF_PREFER flag. This flag should have priority but should not
override a held package.

The messages printer for held and preferred packages are also slightly improved
to include the version number as well as the package name.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>

git-svn-id: http://opkg.googlecode.com/svn/trunk@713 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358