opkg_install: Allow overwrite of files that have no owner
In a package install, the function check_data_file_clashes() checks if a file
that is going to be installed exists and if it belongs to any package.

Currently if a file exists but doesn't belong to any package, the installation
fails as a result of a file clash. In comparison, some other package managers
allow such files to be overwritten silently.

To support both modes of operation, a new option "overwrite_no_owner" is added
which can be set in the config file to cause such files to be overwritten. The
default behaviour is unchanged.

Signed-off-by: Jonh Wendell <jonh.wendell@oiwifi.com.br>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>

git-svn-id: http://opkg.googlecode.com/svn/trunk@720 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358