tests/regress: Added tests for virtual packages
Debian documentations says that if a package PROVIDES and CONFLICTS another
package, it indicates that the provided package is virtual but no other provider
of that virtual package should be installed at the same time. If we attempt to
install two providers of a virtual package 'v', one of which CONFLICTS 'v', it
should not matter which order we try to install them in, the second one should
be blocked. With the patches recently submitted by Andreas Oberritter this works
as expected if first package CONFLICTS 'v'. The current code doesn't work if the
second package CONFLICTS 'v' but the first does not.

Tests are added for both cases, with the latter case being expected to fail.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>

git-svn-id: http://opkg.googlecode.com/svn/trunk@735 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358