autogen.sh: remove the automatic configure call and rewrite file
The automatic start of the configure script after the run of autogen.sh is not
realy useful, in most cases the user want to run configure with self defined
options.

Rewrite autogen.sh to automatic check for the needed tools by autogen.sh and
give some more verbose output. Remove glib-gettextize because currently there
is no i18n at all in the project. This can and should be added later.

The INSTALL file is also updated to clearly indicate the steps required to build
and install opkg.

Signed-off-by: Carsten Schoenert <c.schoenert@gmail.com>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>

git-svn-id: http://opkg.googlecode.com/svn/trunk@742 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358