libopkg: add support for FTP over SSL, implicit and explicit modes.
Explicit mode is supported when setting the ftp_explicit_ssl
configuration option to true.
Both ftp:// and ftps:// schemes are supported when using explicit
mode. Implicit mode is enabled by Curl if ftp_explicit_mode is
not set to true and if the URL starts with "ftps://".
Implicit and explicit modes require curl and SSL support.

Signed-off-by: Julien Gilli <julien.gilli@gmail.com>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>