opkg_verify_cached_file: Fix incorrect return value
This function was incorrectly returning 0 (indicating that the file was already
downloaded completely) when it should have been returning 1 (indicating that the
file required further downloading). When wget was in use this was always
happening and so nothing was getting downloaded.

So the default return value is now 1 at the end of the function and we return 0
earlier in the function where we know that no further downloading is needed.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>