pkg_extract, file_util: Support only gzip compression
For now we will just support the compression methods which were used previously.
The option to use a wider variety of compression methods for package feeds and
packages themselves will be added back at a later date, dependent on a configure
option so that it is not enabled by default.

This should greatly reduce the size of opkg when statically linking against
libarchive.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>