pkg: Add per-package force_reinstall flag
This flag allows two package object to share the same version information but be
distinguishable from each other. It is never stored on disk or reported to the
user but is used when comparing package versions.

A new function, pkg_compare_versions_no_reinstall(), is added to compare
versions without caring about this force_reinstall flag.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>