prepare_url_for_install: Refactor and improve
The heuristics used to determine what the given URL references are improved and
file handling is split out into a separate function.

Reinstalling an existing package version by specifying a filename or URL instead
of the package name is now blocked unless the force_reinstall flag is set as it
will just cause problems.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>