pkg_extract: Canonicalise path before strcmp in extract_file_to_stream
A call the transform_dest_path() is added to remove prefixes such as './' from
the path of an entry in an archive before comparing that to the requested path.
Without the transform, the paths will not be equal even if they refer to the
same file.

This error was reported by paintitgray in the comments on issue #88. The test
package in question had the control file named './control'.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>