pkg_extract: Fix path matching in extract_file_to_stream
I was a little too quick to push the previous fix!

transform_dest_path() returns non-zero for paths we never wish to extract, such
as '.'. If that was encountered before the requested file,
extract_file_to_stream() would return an error rather than skipping it and
continuing to look.

To fix this we just ignore the return value of transform_dest_path() as if a
caller of extract_file_to_stream() explicitly requests something like '.', we
should at least try to fulfil that request.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>