opkg_install: Set pkg state and give advice when install fails
The status line for a package for which installation had failed was the cryptic
"install prefer,user not-installed". This is very unclear and doesn't represent
the broken state of the package clearly. Instead we set flags so that we get
"install reinst-required half-installed". This is better but still not perfect
as 'install' could be misinterpreted. Further cleanup of the state formatting
will improve this.

We also give some advice on how to remove package debris or reattempt the
install.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>