opkg_upgrade_cmd: Only upgrade installed packages
Fixes issue 127.

The manual page says that `opkg upgrade` upgrades already installed packages. If
a package given on the command line is not already installed, an error message
should be raised instead of installing the package. This bad argument is skipped
and other arguments are upgraded if they are installed.

The call to opkg_prepare_url_for_install() is removed as it is not necessary to
handle remote URLs or file names when the packages must already be installed.

The unused variable 'arg' is also removed.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>