libopkg: Deprecate API
The current libopkg API isn't very good and there are many layering violations
between what belongs in the library and what belongs in the frontend program. To
notify users that this API is deprecated and will be removed in v0.4, the API is
only built and installed if '--enable-libopkg-api' is passed to configure.
Passing this argument causes a warning message to be displayed.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>