opkg_verify: Clarify arguments of opkg_verify_file
The file to be verified may not always be a plain text file so calling this
argument 'text_file' could be confusing. Instead the arguments are named to
match the same arguments of opkg_verify_gpg_signature and
opkg_verify_openssl_signature.

The arguments are also given const modifiers as they are not modified.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>