tests/regress: Write MD5Sum to package feed
With the new caching logic, a package without a checksum will always be assumed
to be valid and thus the cached version will always be used. As some of the
tests re-write packages with the same filename in order to test
'--force-reinstall', we need to make sure that the new package file will be used
and not the old one in the cache. Adding an md5sum for each package achieves
this as the md5sum in the package feed will change and thus the cached package
file will be deemed invalid.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>