opkg_verify_signature: Allow method to be selected by config option
The decision as to which signature verification method to use does not have to
be determined at compile time. Instead this decision can be selected at run time
using a configuration option.

If support for the selected method has not been compiled in, an error message is
displayed and the return value is -1.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>