opkg_download_internal: Split out local file handling
This change ensures that local files are handled in the same way whether curl or
wget is used as the download backend and simplifies the logic involved. Although
curl can be used to 'download' local files, it requires an unnecessary amount of
setup to do so.

Currently there is no attempt to check whether the destination file already
matches the source file. If doing so is worthwhile, it can be added.

The function opkg_validate_cached_file is now only needed when curl is used as
the backend, it had no effect for remote files when wget was used as the
backend. The functions create_file_stamp and check_file_stamp are now also only
needed when curl is used as the backend. The function str_starts_with is moved
to simplify the ifdef logic.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>