opkg_download: Rename cleanup function and fix usage
The caller of the cleanup function should not care that curl is currently the
only backend which needs cleanup. The function is renamed to
opkg_download_cleanup and is unconditionally called during exit from opkg or the
library API. The function is also separately defined for each backend to reduce
the use of ifdefs.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>