opkg_download: Split backends into separate files
All code to interact with the curl backend is moved into opkg_download_curl.c
and all code for the wget backend is moved to opkg_download_wget.c. No separate
headers are needed for these files as the definition of opkg_download_backend is
added to opkg_download.h.

Logic is added to the Makefile to ensure the correct file is chosen at compile
time.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>