Cleanup trailing whitespace
The whitespaces has been removed automatically using the following command line:

grep --exclude-dir .git -rHn '[[:space:]]$' | \
	awk -F: '{ print $1 }' | sort | uniq | \
	grep -v COPYING | \
	while read f; do sed -i 's/[[:space:]]*$//g' $f; done

The COPYING file has been spared because (a) it's a GNU file, not an opkg file and
(b) the ^L control chars are in the [[:space:]] character list (and we don't want
them to disapear).

Changes to ChangeLog.ipkg were dropped as they are unnecessary for such a legacy
file.

Signed-off-by: Emmanuel Deloget <logout@free.fr>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>