opkg_conf: Make lists_dir a normal option
The syntax for lists_dir made no sense. It is not a good idea to maintain
backwards compatability with an old config file syntax when the way this
option is used has clearly changed. The new syntax will be:

    option lists_dir <path>

Just like every other option.

It will be simple to write a script which converts the old syntax to the new
syntax and that will ensure users have a safe upgrade path.

This fixes issue 73.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>