opkg_conf: Make lock_file a runtime config option
The ./configure option '--with-opkglockfile' is dropped and OPKGLOCKFILE is no
longer set at compile time. Instead the lock file path may be set using the
lock_file option in a config file as these config files are parsed before the
lock file is used. The default value is now given by OPKG_CONF_DEFAULT_LOCK_FILE
and to comply with the FHS and avoid the requirement to create any directories
which shouldn't already exist it is placed at '/var/run/opkg.lock'.

The static variable lock_file is also moved into the opkg_config structure.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>