opkg_conf: Add status_file option
This option is used to set the status_file_name for each pkg_dest and will be
interpreted as relative to the root_dir of the pkg_dest. It eliminates the need
for the OPKG_STATUS_FILE_SUFFIX symbol and allows more control over where the
status file directory is placed.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>