doc: add and explain checkpatch.pl
Please note that checkpatch.pl is unnafected by the current state of the
code (unless you use option -f) as it verifies patch lines that begins with
a + sign (i.e. lines that are added to the source code). Thus, the current
indentation state is not an issue and checkpatch.pl can be used right now.

Signed-off-by: Emmanuel Deloget <logout@free.fr>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>