checkpatch: indentation uses space instead of tabs
checkpatch.pl used to search for tabs at the beginning of a line. While
this behavior is perfect for the Linux kernel it doesn't match the style
choices that were made for opkg.

This change forces checkpatch.pl to refuse tabs-based indentation. Since
we may (in the future?) revert this choice, the old behavior is kept and
protected by a new --hard-tabs command line option.

Signed-off-by: Emmanuel Deloget <logout@free.fr>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>