opkg_conf: Fix double free and use after free errors
Recent changes added entries in the options array for members of the opkg_config
structure which didn't previously have an associated option. This caused the
memory pointed to by opkg_config->lock_file and opkg_config->lists_dir to be
freed when the options array was de-initialised. In two cases one of these
variables had already been freed but not set to NULL, causing a double free
error at runtime.

In addition, the opkg_config->lock_file variable is needed in order to unlink
the lock file during opkg_conf_deinit. Therefore, this call to unlink must occur
before the options array is de-initialised.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>