pkg_depends: Include recommends in depended_upon_by
This ensures that packages which are recommended by another installed package
are not removed when opkg_config->autoremove is set. As this is currently the
only use of the depended_upon_by array, it should be a safe change to make. If
depended_upon_by is put to other uses in the future, the dependency type of each
entry may need to be checked.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>