opkg_conf: Add and implement 'combine' option
When the option is enabled, installs and upgrades will be combined via calls to
opkg_install_multiple_by_name and opkg_upgrade_multiple_pkgs respectively.

This should resolve the case of issue 124 where multiple packages depend on a
common package with version constraints in such a way that all dependent
packages need to be upgraded together.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>