opkg_install, opkg_upgrade: Mark packages as SW_INSTALL during prepare
In the functions opkg_prepare_install_by_name and opkg_prepare_upgrade_pkg, the
package to be installed was not marked with the correct state_want flag. Thus
when resolving dependencies in a combined install or upgrade operation only the
current package was being considered as all other listed packages were not
marked to be installed.

We fix this by setting state_want to SW_INSTALL for each package to be
installed.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>