opkg_upgrade: Reset state_want flags after failed install
If we attempt to upgrade a package but the installation of the new version of
the package fails, we need to restore the state_want flags to the values they
would have held prior to the failed installation. This ensures that subsequent
installs or upgrades do not consider the old package as scheduled for removal.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>