pkg_hash_fetch_best_installation_candidate: Fix error messages
The error message indicating that packages had been found but were incompatible
with the current architecture was printed even if the reason for packages not
being chosen was not due to architecture incompatibility. This could lead to
confusion and so this message was removed.

If the reason for a package not being selected is architecture incompatibility,
a notice is printed to this effect at an earlier stage in the function.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>