pkg_hash_fetch_best_installation_candidate: Call constraint_fcn earlier
We move the check to see whether a package satisfies the provided constraint_fcn
to an earlier position in the function. It makes more sense to not add a package
to the matching_pkgs list if it fails to satisfy constraint_fcn than it does to
add it to matching_pkgs regardless and then skip over it if it does not satisfy
constraint_fcn.

This has the added advantage of silencing notices that a package would break
existing dependencies or had an incompatible architecture if that package could
not be chosen anyway as it would fail to satisfy constraint_fcn.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>