pkg_hash_fetch_best_installation_candidate: Preserve held pkgs
After recent changes, packages which fail to satisfy constraint_fcn or the
dependencies of packages already installed will not be added to the
matching_pkgs list. This meant that state_flag was never checked to see if any
of these packages were marked with SF_HOLD and so should not be upgraded.
Therefore, if something had occurred which changes the constraints or
dependencies to be satisfied (for example, installing a different package with
--force-depends whose dependencies are not satisfied by the held package), the
next time the held package is considered it could be upgraded despite the hold
flag being set.

To avoid this issue we add all packages which are marked with SF_HOLD to the
matching_pkgs list without checking constraint_fcn, existing dependencies or
package architecture.

The logic to check members of matching_pkgs for SF_HOLD and SF_PREFER are
preserved later in the function and are still needed. This change simply ensures
they will still work as intended.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>