opkg_conf: Refactor lock file handling
The code to create the lock file is factored out of opkg_conf_load into
opkg_lock and the code to destroy the lock file is factored out of
opkg_conf_deinit and the error handling code of opkg_conf_load into opkg_unlock.
This ensures that the unlocking code is consistent between the two places it is
needed and improves readability.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>