pkg_hash_fetch_satisfied_dependencies: Fix constraint
The constraint function used when calling
pkg_hash_fetch_best_installation_candidate should be pkg_constraint_satisfied
not pkg_installed_and_constraint_satisfied.

The constraint function is now treated as a hard constraint on the packages
returned and therefore the previously used constraint function was inappropriate
if we wish to get a list of dependencies to install.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>