pkg_hash_fetch_best_installation_candidate: Restore good_pkg_by_name behaviour
The behaviour around the variable good_pkg_by_name was changed when the call to
constraint_fcn was moved. The previous behaviour (at v0.2.2) was correct when
pkg_hash_fetch_best_installation_candidate was called from
pkg_hash_fetch_best_installation_candidate_by_name as in that case
constraint_fcn would determine whether the package name matched that originally
requested (hence the variable name "good_pkg_by_name"). However, when
pkg_hash_fetch_best_installation_candidate was called from a different function
this behaviour was incorrect as constraint_fcn was intended to be a hard
constraint.

Assuming that the desired behaviour was to set good_pkg_by_name if the package
name matched the orignally requested name, a new check was introduced around the
code which sets good_pkg_by_name. The originally requested name can be assumed
to be apkg->name.

To give an example of the intended behaviour, say we have packages "x" and
"x_alt" where "x_alt" provides "x", and we're trying to find the best
installation candidate for "x". In this case, good_pkg_by_name would be set to
package "x" as it matches the original name, making it preferable to "x_alt".

This behaviour may still not be perfect, but it restores what I assume to be
the intended behaviour of v0.2.2.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>