opkg_conf: Add and implement cache_local_files option
The default behaviour of opkg is changed to allow local files to be linked
rather than copied into the cache directory. However, if the cache_local_files
option is set, the old behaviour will be followed and local files will be copied
into the cache.

A local file is any which is accessed using a 'file://' protocol.

This patch fixes issue 139.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>