opkg_remove_pkg: Remove from_upgrade argument
The from_upgrade argument was never set to a non-zero value when opkg_remove_pkg
was called. Additionally, the argument was called 'message' in the header
instead of 'from_upgrade'.

This confusing mess can just be removed, removal of an old package during an
upgrade is handled by opkg_install_pkg without calling opkg_remove_pkg.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>