opkg_install: Fix level of message if conffile is changed
If a conffile has been changed and force_maintainer is not set, a message should
be printed at the level of NOTICE not at the level of ERROR. The user still
needs to know that there may be changes to the conffile which need to be merged,
but this is not an ERROR as it does not indicate that something wrong has
happened.

This partially fixes issue 128.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>