opkg_conf: Add and implement ignore_maintainer option
If this option is set, the maintainer's changes to a conffile will be ignored
and a message will be printed to indicate that this has happened. This option is
useful when it is known that the user's changes to a conffile should override
those made in an upgrade of a package and allows space to be saved within the
configuration directory.

This completes the fix for issue 128.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>