pkg_depends: Prevent repeats in depended_upon_by
Duplicate entries in the depended_upon_by list may cause a package to not be
considered an orphan when it should be, or cause removals to be attempted twice
when removing all dependent packages. Now that depended_upon_by is a pkg_vec, it
is easy to ensure that duplicates do not enter the list.

This change should fix issue 140.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>