opkg_install: Fix check for orphan dependents
In pkg_remove_orphan_dependent(), we look for DEPENDS or RECOMMENDS which
existed for old_pkg but don't exist for pkg. Previously this was done by
comparing 'struct depend_t' addresses to see if the exact same dependency
existed in both packages. However, each 'struct depend_t' addressed in the
possibilites list is allocated separately during parseDepends(), so two parses
of the exact same dependency string will result in different pointers. Thus we
should not be directly comparing 'struct depend_t' addresses.

Instead, we can compare the addresses of the abstract packages which each
'struct depend_t' references. These addresses are set via a call to
ensure_abstract_pkg_by_name() which will return the same pointer if it sees the
same pkg name several times as the abstract_pkg_t objects are stored in a hash
table.

This change also ensures that if the constraint within a dependency changes
between old_pkg and pkg, it is not considered an orphan. The constraints of the
dependencies of the new package will already have been checked prior to
installing the new package so we know that such changes will have already been
handled.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>