pkg_hash: Fix selection with multiple equal providers
In pkg_hash_fetch_best_installation_candidate, we should not error out if
multiple providers of a given abstract package are present but there is nothing
to differentiate between them (neither is held, preferred, matches the abstract
package name or is already installed). Instead we should just select any valid
provider and get on with the install.

This fixes issue 152.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>