tests/regress: Add test case for issue 152
In the case where multiple packages provide the same abstract package, but there
is nothing to differentiate between the providers (neither is held, preferred,
matches the name of the abstract package or is already installed),
pkg_hash_fetch_best_installation_candidate errors out instead of picking one of
the available providers. This is a vestige of an old code path where opkg
expected the user to explicitly select a provider in this case. Instead we
should now just pick any valid provider.

This bug was found by the OpenEmbedded/Yocto Project autobuilder when testing
opkg v0.2.3. Both 'libperl5' and 'perl' provided 'perl-module-config', but
rootfs creation was failing and opkg said it couldn't find a provider of
'perl-module-config'.

This test case is currently expected to fail. A fix is known and will be merged
shortly.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>