Initial reformat with 'indent'
For each file, the command executed was:
    indent -linux -i4 -nut -sc -bbo -nhnl $f

This is the first step in improving the code formatting and further work is
still required. 'indent' is confused by the 'extern "c" {}' wrapping in the
header files and gets the indentation wrong. It also produces sub-optimal
formatting for some long lines.

Each file will be reviewed by hand after this commit.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>