opkg_download_curl: Refactor calls to curl_easy_setopt
Most calls to curl_easy_setopt in the initialisation function should be checked
for success or failure and a debugging message should be printed if they fail.
For these calls, a setopt macro is introduced which wraps curl_easy_setopt with
error checking and printing of the debug message as needed. This macro is now
used for all calls to curl_easy_setopt other than those which need to print an
error message and cause the initialisation function to abort if they fail.

This change adds error checking to all curl_easy_setopt calls in the
initialisation function which were previously missing error checking.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>