opkg_install: Improve logic in satisfy_dependencies_for
If opkg_config->force_depends is set, the old code would never set 'conflicts',
therefore 'if (conflicts)' would be false and the bulk of the code in this
function would not be executed. This meant that the code to set 'level' based on
the value of opkg_config->force_depends was nonsensical.

This situation is improved by simply returning 0 early of
opkg_config->force_depends is set.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>