opkg_install: Fix segfault with force_reinstall set
If a package is reinstalled with force_reinstall set, the 'new' version of the
package used for the reinstall will be initialised by reading the package file
itself. Therefore, this pkg object will not have a valid pkg->src. Calling
pkg_src_verify(NULL) results in a segfault.

This issue is solved by checking that the package has a valid pkg->src before
trying to verify it. This should also ensure that the installation of packages
directly from a local package file will succeed.

For reinstallations this does not present a security risk as the original
package feed and file will be verified during opkg_prepare_url_for_install
before this file is used to construct the new pkg object.

This fixes issue 153.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>