opkg_install: Fix error status during upgrade
During an upgrade, if A is upgraded, and A depends on B, B is upgraded
automatically. Then the upgrade procedures moves to B, where it fails,
as B was already updated (issue 154).

Skip package if the version to be installed and the version on the
system match AND it's an upgrade operation.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>