opkg: Add --host-cache-dir command line argument
This argument tells opkg to store the cache directory outside the offline root
directory if the -o option is passed. It may be used to keep the offline root
directory clean of irrelevant cache files. It may also be helpful if the offline
root path is very long and placing cache files under this directory has a risk
of exceeding PATH_MAX.

Eg.
    'opkg --cache-dir /foo -o /offline ...' stores the cache in '/offline/foo'.

    'opkg --cache-dir /foo -o /offline --host-cache-dir ...' stores the cache in
        '/foo'.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>