opkg_download.c: fix memory leak on opkg_download_pkg
pkg->local_filename is being reassigned to a new string (without freeing
the old one) when a package is not present on the cache. Make sure
pkg->local_filename is not reassigned by calling opkg_download_internal
directly, as there is no need to recompute the value of the local cache
filepath.

Fixes issue 8517

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>