pkg_get_provider_replacees: do not add installed pkg to replacee list
If package A replaces provider B, and B is provided by A,
pkg_get_provider_replacees incorrectly adds A to the list of B replacees
when A is installed. During an upgrade, pacakge A is removed during
pkg_remove_installed_replacees, then once more during the package
upgrade.

Add check to skip the insertion of package A into the replacees vector
in pkg_get_provider_replacees.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>