libopkg: add download_first option
When enabled, opkg will download all the packages needed by the
install/upgrade operation before executing the transaction. Adds an
extra layer of robustness in case of network outages on operations that
take a long time (for example, during upgrades).

This option is not available in the internal solver backend since
multiple operations are not combined on a single solve. Instead, there
is an individual solve for each transaction. This is true even if the
combine flag is set.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>