libopkg/opkg_message.c: Set freed pointer to NULL.
This allows us to "clear" the list of errors as opposed to just
freeing them, e.g. when the application terminates. This makes
sense if you are, for example, developing an update-manager that
is continously running in the background and every now and then
performs an action, e.g. update, upgrade, etc., but you still
wish to report errors for each action, not all errors that
occured since the application started.

Signed-off-by: Simon Aittamaa <simon.aittamaa@gmail.com>