__pkg_hash_fetch_conflictees: fix conflict logic
The condition to detect a new conflict makes sure that the installed
package being analyzed does not provide the pkg being installed. It is
possible for a pkg to Provides/Conflicts another one, in which case the
conflict will not be reported. Fix condition to compare pkg names to
only skip the conflict if the pkg names are the same (a pkg shouldn't
conflict with itself).

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>