libopkg: kill active_list
The active_list abstraction has a long history in opkg.  It's purpose
was to be a richer list type which, given information about a dependency
relationship, can use it to control order of traversal.

However, the "richness" of this data type has never been used.  In fact,
as of commit d0c00a26eb ("active_list_add_depend: remove unused
function"), the interface used to specify dependencies was removed.

Get rid of the abstraction entirely by converting users to using the
basic list type.

Signed-off-by: Josh Cartwright <joshc@ni.com>
Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>