libopkg/opkg_gpg.c: Fix to gpg signature verification.
Previously the GPG check would simply perform a digest against a give repo/package
from the signature file. At no point would it verify the package's signature was
signed properly. This fix verifies the package/repository not only checks the digest
but that the package is also signed by a trusted source.

Signed-off-by: Bryan Krause <bryan.krause@baesystems.com>
Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>