libopkg: add Installed-Size field to status file
While opkg maintained a pkg->installed_size variable, it was not using
it to either store or print installed size information on the opkg
commands info and status.

By capturing the installed size at extract time, now opkg can store (and
display) the Installed-Size field on the status file.

Fixes 9906

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>