libopkg: add per-feed signature verification toggle
To allow per-feed signature verification control, the format for feeds
is extended to allow options. The options format is the same one used
by apt:

src{/gz] name uri [option=value option=value]

For now, only the trusted option is supported, but in the future we may
add other options.

trusted: the value yes tells opkg always to consider this source as trusted, even
if it doesn't pass authentication checks. Default is no.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>