populate_installed_repo: make autoinstalled packages removable
In a system that has package 'a' & 'b' installed, where package 'a'
recommends 'b', trying to install package 'c', which depends on 'd'
(which conflicts with 'b'), the expected behavior is for the package
manager to uninstall the recommended package 'b', then proceed to
install 'c' & 'd'. By making autoinstalled packages potentially removable,
libsolv will correctly remove installed recommended packages that
conflict with a to be installed package (or any of it's dependencies).

Fixes bugzilla 10777 (libsolv only)

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>